Added Perm
The solution to the "Perm" problem, namely "Enumerating Gene Orders", has been added. This lead to the creation of a new module, `Solidran.Math` that contains math related utilities (only the `factorial` function for now), which are, for example, used in this solution.

There was no need for a `Detail` module. The solution is so trivial with Haskell that only 1 "home-made" function was used (not even needed) and is tested in `Solidran.MathSpec`.